#6 Refactor proxy authentication
